requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
langgraph
faiss-cpu
openai
python-dotenv
pydantic>=2.10
tiktoken
pypdf
python-docx
sqlalchemy
numpy
